+++ src/JavLuv/MovieDetailViewModel.cs
@@ -389,6 +389,8 @@ namespace JavLuv
                     m_movieData.Metadata.Actors = actors;
                     m_movieData.MetadataChanged = true;
                     NotifyPropertyChanged("Actors");
+                    Actresses = new MovieDetailActressViewModel(this);
+                    NotifyPropertyChanged("Actresses");
                 }
             }
         }